Separate all methods in a new file
