+++ scripts/src/java/org/oppia/android/scripts/common/BUILD.bazel
@@ -46,3 +46,28 @@ kt_jvm_library(
     srcs = ["RepositoryFile.kt"],
     visibility = ["//scripts:oppia_script_library_visibility"],
 )
+
+kt_jvm_library(
+    name = "license_fetcher",
+    testonly = True,
+    srcs = [
+        "LicenseFetcher.kt",
+        "LicenseFetcherImpl.kt",
+    ],
+    visibility = ["//scripts:oppia_script_library_visibility"],
+)
+
+kt_jvm_library(
+    name = "maven_dependencies_list_generator",
+    testonly = True,
+    srcs = ["MavenDependenciesListGenerator.kt"],
+    visibility = ["//scripts:oppia_script_library_visibility"],
+    deps = [
+        ":bazel_client",
+        "//scripts/src/java/org/oppia/android/scripts/common:license_fetcher",
+        "//scripts/src/java/org/oppia/android/scripts/maven/data:data_lib",
+        "//scripts/src/java/org/oppia/android/scripts/proto:maven_dependencies_java_proto",
+        "//scripts/src/java/org/oppia/android/scripts/proto:maven_dependencies_java_proto_lite",
+        "//third_party:com_google_protobuf_protobuf-java",
+    ],
+)
