+++ scripts/src/java/org/oppia/android/scripts/common/LicenseFetcher.kt
