+++ scripts/src/java/org/oppia/android/scripts/common/LicenseFetcherImpl.kt
