+++ scripts/src/java/org/oppia/android/scripts/maven/BUILD.bazel
@@ -5,22 +5,12 @@ Libraries corresponding to attributing third-party library copyright licenses in
 load("@io_bazel_rules_kotlin//kotlin:kotlin.bzl", "kt_jvm_library")
 
 kt_jvm_library(
-    name = "license_fetcher",
-    testonly = True,
-    srcs = [
-        "LicenseFetcher.kt",
-        "LicenseFetcherImpl.kt",
-    ],
-)
-
-kt_jvm_library(
     name = "generate_maven_dependencies_list_lib",
     testonly = True,
     srcs = ["GenerateMavenDependenciesList.kt"],
-    visibility = ["//scripts:oppia_script_binary_visibility"],
+    visibility = ["//visibility:public"],
     deps = [
-        ":license_fetcher",
-        "//scripts/src/java/org/oppia/android/scripts/common:bazel_client",
+        "//scripts/src/java/org/oppia/android/scripts/common:maven_dependencies_list_generator",
         "//scripts/src/java/org/oppia/android/scripts/maven/data:data_lib",
         "//scripts/src/java/org/oppia/android/scripts/proto:maven_dependencies_java_proto",
         "//scripts/src/java/org/oppia/android/scripts/proto:maven_dependencies_java_proto_lite",
